add update_available flag for cached_models, and moved /update endpoint from inference -> models
