+++ studio/backend/models/inference.py
@@ -117,20 +117,6 @@ class UnloadRequest(BaseModel):
     model_path: str = Field(..., description = "Model identifier to unload")
 
 
-class UpdateRequest(BaseModel):
-    """Request to update a model"""
-    model_path: str = Field(..., description = "Model identifier or local path to update")
-    hf_token: Optional[str] = Field(
-        None, description = "HuggingFace token for gated models"
-    )
-    has_vision: Optional[bool] = Field(
-        False, description = "Whether this model has vision capabilities"
-    )
-    gguf_variant: Optional[str] = Field(
-        None, description = "GGUF quantization variant (e.g. 'Q4_K_M')"
-    )
-
-
 class ValidateModelRequest(BaseModel):
     """Check whether an identifier resolves to a ModelConfig; does NOT load weights."""
 
@@ -195,11 +181,6 @@ class GenerateRequest(BaseModel):
     repetition_penalty: float = Field(1.0, ge = 1.0, le = 2.0, description = "Repetition penalty")
     presence_penalty: float = Field(0.0, ge = 0.0, le = 2.0, description = "Presence penalty")
     image_base64: Optional[str] = Field(None, description = "Base64 encoded image for vision models")
-
-
-class UpdateResponse(BaseModel):
-    """Response after updating a model"""
-    model_path: str = Field(..., description = "Updated model path")
 
 
 class LoadResponse(BaseModel):
