+++ studio/backend/routes/inference.py
@@ -1052,12 +1052,10 @@ async def _aiter_llama_stream_items(
 from models.inference import (
     LoadRequest,
     UnloadRequest,
-    UpdateRequest,
     GenerateRequest,
     LoadResponse,
     LoadProgressResponse,
     UnloadResponse,
-    UpdateResponse,
     InferenceStatusResponse,
     ChatCompletionRequest,
     ChatCompletionChunk,
@@ -2493,65 +2491,6 @@ def _maybe_unsupported_message(msg: str) -> str:
     if any(h.lower() in msg.lower() for h in _NOT_SUPPORTED_HINTS):
         return f"This model is not supported yet. Try a different model. (Original error: {msg})"
     return msg
-
-
-@router.post("/update", response_model = UpdateResponse)
-async def update_hf_model(
-    request: UpdateRequest,
-    fastapi_request: Request,
-    current_subject: str = Depends(get_current_subject),
-):
-    try:
-        llama_backend = get_llama_cpp_backend()
-        config = ModelConfig.from_identifier(
-            model_id = request.model_path,
-            hf_token = request.hf_token,
-            gguf_variant = request.gguf_variant,
-        )
-        if not config:
-            raise HTTPException(
-                status_code = 400,
-                detail = f"Invalid model identifier: {request.model_path}",
-            )
-        if config.is_local:
-            raise HTTPException(
-                status_code = 400,
-                detail = "Only Hugging Face models can be updated.",
-            )
-
-        if config.is_gguf:
-            if not config.gguf_hf_repo:
-                raise HTTPException(
-                    status_code = 400,
-                    detail = "GGUF update requires a Hugging Face repo.",
-                )
-            model_path = await asyncio.to_thread(
-                llama_backend._download_gguf,
-                hf_repo = config.gguf_hf_repo,
-                hf_variant = config.gguf_variant,
-                hf_token = request.hf_token,
-            )
-            if config.is_vision:
-                await asyncio.to_thread(
-                    llama_backend._download_mmproj,
-                    hf_repo = config.gguf_hf_repo,
-                    hf_token = request.hf_token,
-                )
-        else:
-            from huggingface_hub import snapshot_download
-
-            hf_repo = config.base_model if config.is_lora and config.base_model else config.path
-            model_path = await asyncio.to_thread(
-                snapshot_download,
-                repo_id = hf_repo,
-                token = request.hf_token,
-            )
-        return UpdateResponse(model_path = model_path)
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Error updating model '{request.model_path}': {e}", exc_info = True)
-        raise HTTPException(status_code = 500, detail = f"Failed to update model: {str(e)}")
 
 
 @router.post("/load", response_model = LoadResponse)
